Clean-up logic of working with platform-specific markers in AbstractCodeMetaInfoTest
Note that if code meta info has platform-attribute which is not equal to
the current host-platform, then it is *ignored* (makes sense, because we
can't verify it on the current host).

The current code though treats presence of ANY attribute as a presennce
of platform-attribute. Which means that if someone introduces another
non-platform attribute and uses codemetainfo with such attribute, that
CMI will be *ignored forever*.

This is what actually happened in commonSealedPlatformWithInheritor,
which somehow got a marker with {JVM} attribute, and this marker has
never checked and was just written to actual testdata.

This commit makes checking for platform-attributes more robust and
explicit, which fixes the testdata in that test