Add API for getting the original artifact from `KeepOriginalDependency`
This API is needed for IDE import implementation for KPM, which aims to
avoid using Gradle `Configuration`s directly for getting the artifacts
and therefore needs the API in KPM dependency resolvers for getting the
artifacts of all dependencies, including the "plain" dependencies that
KPM resolvers don't interpret in any special way (e.g. by decomposing
the composite artifacts).